% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_realizations.R
\name{make_realizations}
\alias{make_realizations}
\title{make_realizations}
\usage{
make_realizations(predObject, nreals = 200)
}
\arguments{
\item{predObject}{the object output from predict_hetgp()}

\item{nreals}{number of realizations to generate (default is 200)}
}
\value{
a list containing a data.frame of realizations and a matrix containing realizations
}
\description{
make_realizations
}
\examples{
data(sample_lake_data_1mdepth)
mod1 =  fit_hetgp(X = "DOY", Y = "temperature",site_id = "FCR", df = sample_lake_data_1mdepth)
preds <- predict_hetgp(het_gp_object = mod1, reference_datetime = as.Date("2022-10-05"), save_covmat=TRUE)

sims = make_realizations(predObject = preds)

}
