% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_hetgp.R
\name{plot_hetGPpreds}
\alias{plot_hetGPpreds}
\title{Plot predictive mean and prediction intervals of a forecast}
\usage{
plot_hetGPpreds(predObject)
}
\arguments{
\item{predObject}{an object of class "hetGPpreds"}
}
\value{
it makes a plot
}
\description{
Plot predictive mean and prediction intervals of a forecast
}
\examples{
data(sample_lake_data_1mdepth)
mod1 =  fit_hetgp(X = "DOY", Y = "temperature", site_id = "FCR", df = sample_lake_data_1mdepth)
preds = predict_hetgp(het_gp_object = mod1, reference_datetime = "2023-09-01")
plot_hetGPpreds(predObject=preds)

}
