% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hetgp.R
\name{fit_hetgp}
\alias{fit_hetgp}
\title{Make predictions from a heteroscedastic Gaussian process model fit}
\usage{
predict_hetgp(het_gp_object,
                          save_covmat = FALSE,
                          reference_datetime,
                          max_horizon = 35,
                          depths = 1:10,
                          PI = .90)
}
\arguments{
\item{het_gp_object}{the fit object from fit_hetgp()}

\item{save_covmat}{boolean value denoting whether to save the covariance matrix at new input values (default is FALSE)}

\item{reference_datetime}{the reference datetime. Must be of class string, POSIXct, or Date and formatted as YYYY-MM-DD.}

\item{max_horizon}{maximum forecast horizon in days (default is 35)}

\item{depths}{Depths in m at which predictions are desired (default is 1:10)}
\item{PI}{Width of prediction intervals (numeric value between 0 and 1)}
}
\value{
an object of class "hetGPpreds" containing the prediction data.frame in standard format, the covariance matrix,
the original data.frame used for model fitting, a data.frame containing means, standard deviations and upper and lower prediction bounds at prediction locations for easy plotting, a boolean value denoting whether depth was used as a covariat during model fitting, and the prediction interval width
}
\description{
Obtain the posterior predictive mean and covariance from a heteroscedastic Gaussian process fit at new input locations
}
\examples{
preds <- predict_hetgp(het_gp_object = het_object,
                          save_covmat = FALSE,
                          reference_datetime = "2023-09-19",
                          max_horizon = 35,
                          PI = .90)
}

