% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_realizations.R
\name{rmultnorm}
\alias{rmultnorm}
\title{Sample from a multivariate normal distribution}
\usage{
rmultnorm(n, mu, sigma)
}
\arguments{
\item{n}{number of realizations}

\item{mu}{mean vector of length m}

\item{sigma}{covariance matrix of dimension mxm}
}
\value{
a matrix of dimension n x m
}
\description{
Sample from a multivariate normal distribution
}
\examples{
covpow <- function(locs,pow=2,scale=5){
d1 <- dist(locs)
n <- dim(locs)[1]
mat <- matrix(0,n,n)
mat[lower.tri(mat)] <- d1
mat <- mat+t(mat)
cc <- exp(-(mat/scale)^pow)
return(cc)
}
locs <- expand.grid(1:20,1:20)
locs <- cbind(locs[,1],locs[,2])
covmat<- covpow(locs)
z <- rmultnorm(1,rep(0,20*20), covmat)
}
