% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_lake_data_1mdepth}
\alias{sample_lake_data_1mdepth}
\title{Lake temperature data from Falling Creek Reservoir at 1m depth}
\format{
## `sample_lake_data_1mdepth`
A data frame with 1795 rows and 4 columns:
\describe{
  \item{datetime}{date formatted as YYYY-MM-DD}
  \item{site_id}{Name of reservoir}
  \item{variable}{name of variable}
  \item{observation}{recorded data}
}
}
\source{
<https://s3.flare-forecast.org/targets/fcre_v2/fcre/fcre-targets-insitu.csv>
}
\usage{
sample_lake_data_1mdepth
}
\description{
A subset of sensor data from Falling Creek Reservoir, Virginia, USA
}
\keyword{datasets}
