% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_hetgp.R
\name{predict_hetgp}
\alias{predict_hetgp}
\title{predict_hetgp}
\usage{
predict_hetgp(het_gp_object, save_covmat = FALSE)
}
\arguments{
\item{save_covmat}{boolean: should the predictive covariance matrix between prediction locations be saved?}

\item{het_gp_fit}{a hetGP model fit object (from fit_hetgp())}
}
\value{
a data.frame in standard format
}
\description{
predict_hetgp
}
\examples{
preds <- predict_hetgp(het_gp_object = het_object)

}
