% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_hetgp.R
\name{predict_hetgp}
\alias{predict_hetgp}
\title{Obtain posterior predictive distribution from a heteroscedastic Gaussian process model fit}
\usage{
predict_hetgp(
  het_gp_object,
  save_covmat = FALSE,
  reference_datetime,
  max_horizon = 35,
  depths = 1:10,
  PI = 0.9
)
}
\arguments{
\item{het_gp_object}{a hetGP model fit object (from fit_hetgp())}

\item{save_covmat}{boolean: should the predictive covariance matrix between prediction locations be saved?}

\item{reference_datetime}{character or Date, POSIXt class date formatted as YYYY-MM-DD}

\item{max_horizon}{forecast horizon in days (default is 35)}

\item{depths}{depths at which predictions are desired (numeric vector >= 0) (default is 1:10)}

\item{PI}{prediction interval coverage (numeric between 0 and 1) (default is 1:10)}
}
\value{
a list containing a data.frame in standard format containing forecasts; a predictive covariance matrix if save_covmat = TRUE (otherwise this will be NULL),
the original df used for fitting, a data.frame containing the mean, sd and upper and lower bounds (easier for plotting);
a boolean value denoting whether depth was a covariate, and the name of the response variable
}
\description{
Obtain posterior predictive distribution from a heteroscedastic Gaussian process model fit
}
\examples{
data(sample_lake_data_1mdepth)
mod1 = fit_hetgp(X = "DOY", Y = "temperature",site_id = "FCR", df = sample_lake_data_1mdepth)
preds <- predict_hetgp(het_gp_object = mod1, reference_datetime = as.Date("2022-10-05"))
}
\references{
Binois, Mickael, and Robert B. Gramacy. "hetgp: Heteroskedastic Gaussian process modeling and sequential design in R." (2021).
}
